Refactor adapter files and update include paths in loc_map module
