+++ src/loc_map/include/adapter_loc_map/adapter.hpp
