+++ src/loc_map/src/loc_map/lm_node.cpp
@@ -1,8 +1,8 @@
 #include "loc_map/lm_node.hpp"
 
-#include "adapter/eufs.hpp"
-#include "adapter/fsds.hpp"
-#include "adapter/map.hpp"
+#include "adapter_loc_map/eufs.hpp"
+#include "adapter_loc_map/fsds.hpp"
+#include "adapter_loc_map/map.hpp"
 #include "utils/car.hpp"
 #include "utils/formulas.hpp"
